feat/rename-progress-components
