+++ packages/components/src/index.ts
@@ -82,12 +82,12 @@ export { default as Loader } from './loader/Loader';
 export * from './loader/Loader';
 export { default as MediaObject } from './mediaObject/MediaObject';
 export * from './mediaObject/MediaObject';
-export { default as Path } from './progressIndicator/Path';
-export * from './progressIndicator/Path';
+export { default as Path } from './progressCircular/Path';
+export * from './progressCircular/Path';
 export {
-  default as ProgressIndicator,
-} from './progressIndicator/ProgressIndicator';
-export * from './progressIndicator/ProgressIndicator';
+  default as ProgressCircular,
+} from './progressCircular/ProgressCircular';
+export * from './progressCircular/ProgressCircular';
 export { default as Reel } from './reel/Reel';
 export * from './reel/Reel';
 export { default as Select } from './select/Select';
@@ -112,10 +112,8 @@ export { default as Input } from './textInput/Input';
 export * from './textInput/Input';
 export { default as TextInput } from './textInput/TextInput';
 export * from './textInput/TextInput';
-export {
-  default as TextProgressIndicator,
-} from './textProgressIndicator/TextProgressIndicator';
-export * from './textProgressIndicator/TextProgressIndicator';
+export { default as ProgressLinear } from './progressLinear/ProgressLinear';
+export * from './progressLinear/ProgressLinear';
 export { default as Toggle } from './toggle/Toggle';
 export * from './toggle/Toggle';
 export { default as Tooltip } from './tooltip/Tooltip';
