+++ packages/components/src/progressCircular/README.mdx
@@ -1,10 +1,10 @@
 ---
-name: Progress Indicator
+name: Progress Circular
 menu: Components
-route: /components/progress-indicator
+route: /components/progress-circular
 ---
 
-# Progress Indicator
+# Progress Circular
 
 [JIRA link](https://coingaming.atlassian.net/wiki/spaces/CDS/pages/2033778811/Progress+indicator)
 
@@ -17,9 +17,9 @@ Use this when the process has a known duration
 Circular progress indicators display progress by animating an indicator along an invisible circular track in a clockwise direction.
 
 ```jsx react-live
-<ProgressIndicator progress={30} fontSize="5rem" />
+<ProgressCircular value={30} fontSize="5rem" />
 ```
 
 ## See also
 
-If you need a [text progress indicator](/components/text-progress-indicator)
+If you need a [linear progress indicator](/components/progress-linear)
