+++ packages/components/src/progressLinear/README.mdx
@@ -1,10 +1,10 @@
 ---
-name: Text Progress Indicator
+name: Progress Linear
 menu: Components
-route: /components/text-progress-indicator
+route: /components/progress-linear
 ---
 
-# Text Progress Indicator
+# Progress Linear
 
 [JIRA link](https://coingaming.atlassian.net/wiki/spaces/CDS/pages/2033778811/Progress+indicator)
 
@@ -13,8 +13,8 @@ A progress indicator is a visual representation of a user’s progress through a
 ## Usage
 
 ```jsx react-live
-<TextProgressIndicator
-  progress={30}
+<ProgressLinear
+  value={30}
   text="Progress"
   backgroundColor="hit.100"
 />
@@ -22,4 +22,4 @@ A progress indicator is a visual representation of a user’s progress through a
 
 ## See also
 
-If you need a [circular progress indicator](/components/progress-indicator)
+If you need a [circular progress indicator](/components/progress-circular)
